+++ test/wal-config-test.el
@@ -192,30 +192,30 @@
 (ert-deftest wal-tangle-config-prompt ()
   (bydi ((:always wal-tangle-do-prompt)
          (:always y-or-n-p )
-         wal-prelude-tangle-config)
+         wal-tangle-config)
 
     (let ((wal-tangle-do-prompt t))
 
       (wal-tangle-config-prompt)
-      (bydi-was-called wal-prelude-tangle-config))))
+      (bydi-was-called wal-tangle-config))))
 
 (ert-deftest wal-tangle-config-prompt--after ()
   (bydi ((:always wal-tangle-do-prompt)
          (y-or-n-p . #'ignore)
-         wal-prelude-tangle-config
+         wal-tangle-config
          message)
 
     (let ((wal-tangle-do-prompt t))
 
       (wal-tangle-config-prompt)
 
-      (bydi-was-called-with message "To tangle, call `wal-prelude-tangle-config'")
+      (bydi-was-called-with message "To tangle, call `wal-tangle-config'")
       (should-not wal-tangle-do-prompt)
 
       (bydi-clear-mocks)
       (wal-tangle-config-prompt)
 
-      (bydi-was-called-with message "Config changed. To tangle, call `wal-prelude-tangle-config'"))))
+      (bydi-was-called-with message "Config changed. To tangle, call `wal-tangle-config'"))))
 
 (ert-deftest wal-config-switch-project ()
   (defvar wal-emacs-config-default-path)